Remove PRId64 (fails on windows builds), now casts int64_t to int Improved RPC logic

+++ src/wallet.cpp
@@ -5258,9 +5258,9 @@ int64_t CWallet::EstimateAnonFeeIncluded(int64_t nMaxAmount, int nRingSize, std:
     }
 
     if (fDebugRingSig){
-        LogPrintf("EstimateAnonFeeIncluded: Found value = %" PRId64 "\n", nValue);
-        LogPrintf("EstimateAnonFeeIncluded: Fee = %" PRId64 "\n", nFeeRet);
-        LogPrintf("EstimateAnonFeeIncluded: Dust = %" PRId64 "\n", GetShadowBalance() - nValue - nFeeRet);
+        LogPrintf("EstimateAnonFeeIncluded: Found value = %i\n", (int) nValue);
+        LogPrintf("EstimateAnonFeeIncluded: Fee = %i\n", (int) nFeeRet);
+        LogPrintf("EstimateAnonFeeIncluded: Dust = %i\n", (int) (GetShadowBalance() - nValue - nFeeRet));
     }
